+++ CHANGES.rst
@@ -1,8 +1,8 @@
 Python Liquid Change Log
 ========================
 
-Version 1.8.0 (unreleased)
---------------------------
+Version 1.8.0
+-------------
 
 **Features**